+++ core/state/frozenfunds/frozen_funds.go
@@ -113,7 +113,7 @@ func (f *FrozenFunds) PunishFrozenFundsWithAddress(fromHeight uint64, toHeight u
 				f.bus.Events().AddEvent(uint32(fromHeight), &eventsdb.SlashEvent{
 					Address:         item.Address,
 					Amount:          slashed.String(),
-					Coin:            item.Coin,
+					Coin:            uint64(item.Coin),
 					ValidatorPubKey: *item.CandidateKey,
 				})
